Fix setPWM not working above 255 channels

+++ Adafruit_TLC59711.cpp
@@ -116,7 +116,7 @@ void Adafruit_TLC59711::write() {
  *  @param  pwm
  *          pwm value
  */
-void Adafruit_TLC59711::setPWM(uint8_t chan, uint16_t pwm) {
+void Adafruit_TLC59711::setPWM(uint16_t chan, uint16_t pwm) {
   if (chan > 12 * numdrivers)
     return;
   pwmbuffer[chan] = pwm;
